modified stopcontainer.sh

+++ stop_container.sh
@@ -2,5 +2,7 @@
 set -e
 
 # Stop the running container (if any)
-containerid = docker ps | awk -f " " '{print $1}'
-docker rm -f $containerid
+containerid=$(docker ps -q)
+if [ -n "$containerid" ]; then
+    docker rm -f $containerid
+fi
